refactor: move ledger-connection components to shared namespace
