+++ apps/next/src/components/ConnectLedger/LedgerConnector/hooks/useLedgerPublicKeyFetcher/index.ts
