+++ apps/next/src/components/ConnectLedger/LedgerConnector/hooks/useLedgerPublicKeyFetcher/useLedgerBasePublicKeyFetcher.ts
