+++ apps/next/src/components/ConnectLedger/LedgerConnector/util/buildAddressPublicKey.ts
