+++ apps/next/src/components/ConnectLedger/LedgerConnector/util/buildExtendedPublicKey.ts
