+++ apps/next/src/components/ConnectLedger/LedgerConnector/util/index.ts
