+++ apps/next/src/components/ConnectLedger/Troubleshooting/InfoBox.tsx
