+++ apps/next/src/components/ConnectLedger/Troubleshooting/Troubleshooting.tsx
