+++ apps/next/src/components/ConnectLedger/images/SolanaGlow.png
